requirements.txt
faiss-cpu
gradio
huggingface-hub
imutils
langchain
langchain-community
langchain-core
langchain-openai
llama-cloud
llama-cloud-services
llama-index
llama-parse
nltk
numpy
openai
opencv-python-headless
pandas
pdfplumber
pillow
PyMuPDF
pytesseract
python-dotenv
requests
sentence-transformers
tqdm
transformers









